backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
alembic==1.13.1

# Google Cloud Vision
google-cloud-vision==3.5.0
google-auth==2.25.2

# Image Processing
Pillow==10.1.0
pdf2image==1.16.3
PyPDF2==3.0.1

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9

# Duplicate Detection
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
phonenumbers==8.13.27
requests==2.31.0

# Development
pytest==7.4.3
pytest-flask==1.3.0
black==23.12.1
flake8==6.1.0
